move testDoubleColumnWithoutDoubleStatistics to TestRecordReaderImpl
